brandings fixed

+++ app/Modules/admin/Controllers/BrandingController.php
@@ -7,7 +7,9 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Input;
 use Illuminate\Support\Facades\Redirect;
 use App\Modules\Admin\Models\Branding;
-
+use Carbon\Carbon;
+use Illuminate\Support\Facades\Validator;
+use Image;
 class BrandingController extends Controller {
 
 
@@ -18,10 +20,10 @@ class BrandingController extends Controller {
 	 * @return Response
 	 */
 	public function index()
-	{	
+	{
 
-        
-        $InstitutionObj = new Institution();
+
+		$InstitutionObj = new Institution();
 		$inst_arr = $InstitutionObj->getInstitutions();
 		/*$inst_arr=Institution::get();*/
 		return view('admin::branding.brand',compact('inst_arr'));
@@ -32,42 +34,83 @@ class BrandingController extends Controller {
 	 *
 	 * @return Response
 	 */
-	public function create()
+	public function create( Request $request)
+
 	{
-	
-		$post=Input::All(); 
+
+		$post = Input::All();
+		$messages = [
+			'institution_id.required' => ' Institution Name  is required',
+			//'title.required' => 'Enter Name of the Title',
+			'hbcolor.required' => 'BackGround color is required  ',
+			'headertc.required' => ' HeadTextColor is required',
+			'boxhbc.required' => ' BoxTextColor is required',
+			'boxhtcolor.required' => ' BoxHeaderColor is required',
+			'btextc.required' => ' BoxTextColor is required',
+			//'buttonc.required' => ' ButtonColor is required',
+		];
+		$rules = [
+			'institution_id' => 'required',
+			//'title'=>'required|unique:title',
+			'hbcolor' => 'required',
+			'headertc' => 'required',
+			'boxhbc' => 'required',
+			'boxhtcolor' => 'required',
+			'btextc' => 'required',
+			//'buttonc' => 'required',
+		];
+		$validator = Validator::make($post, $rules);
+		if ($validator->fails()) {
+			return Redirect::back()->withInput()->withErrors($validator);
+		}
+		if($request->hasFile('image')) {
+			// $file = Input::file('image');
+			// $timestamp = str_replace([' ', ':'], '-', Carbon::now()->toDateTimeString());
+			// $name = $timestamp . '-' . $file->getClientOriginalName();
+			// $image->filePath = $name;
+			// $file->move(public_path() . '/brandingimages/', $name);
+			//dd($file);
+
+
+			$file = Input::file('image');
+			$name = time() . '.' . $file->getClientOriginalExtension();
+
+			$path = public_path('/brandingimages/' . $name);
+
+
+			Image::make($file->getRealPath())->resize(200, 200)->save($path);
+		}
 		$branding = Branding::create([
-			
-			'title'=>$post['title'],
-			/*'filepath'=>$post['filepath'],*/
-			'header_bg_color'=>$post['hbc'],
-			'header_text_color'=>$post['htc'],
-			'box_header_bg_color'=>$post['bhbc'],
-			'box_header_text_color'=>$post['bhtc'],
-			'box_text_color'=>$post['btc'],
-			'button_color'=>$post['bc'],
-			'institution_id'=> $post['institution_id'],
-			]);//dd($branding);
-		$institution_name=Institution::select('name')->where('id',$branding->institution_id)->first();
-		//dd($institution_name);
+			//'title' => $post['title'],
+			'filepath' => $name,
+			'header_bg_color' => $post['hbcolor'],
+			'header_text_color' => $post['headertc'],
+			'box_header_bg_color' => $post['boxhbc'],
+			'box_header_text_color' => $post['boxhtcolor'],
+			'box_text_color' => $post['btextc'],
+			'button_color' => $post['buttonc'],
+			'institution_id' => $post['institution_id'],
+		]);
+		$institution_name = Institution::select('name')->where('id', $branding->institution_id)->first();
 		\Session::flash('success', 'Successfully added.');
-			return Redirect::route('brandview');
-				//return view('admin::branding.brandview',compact('branding','institution_name'));
+		return Redirect::route('brandview');
+
+
 	}
 
-		
+
 
 	/**
 	 * Store a newly created resource in storage.
 	 *
 	 * @return Response
 	 */
-	
-	public function store(Request $request)
+
+	/*public function store(Request $request)
 	{
         $image = new Branding();
         $this->validate($request, [
-            'title' => 'required',
+//            'title' => 'required',
             'image' => 'required'
         ]);
         $image->title = $request->title;
@@ -76,16 +119,16 @@ class BrandingController extends Controller {
             $file = Input::file('image');
             //getting timestamp
             $timestamp = str_replace([' ', ':'], '-', Carbon::now()->toDateTimeString());
-            
+
             $name = $timestamp. '-' .$file->getClientOriginalName();
-            
+
             $image->filePath = $name;
 
             $file->move(public_path().'/images/', $name);
         }
         $image->save();
         return $this->create()->with('success', 'Image Uploaded Successfully');
-	}
+	}*/
 	/**
 	 * Display the specified resource.
 	 *
@@ -94,9 +137,13 @@ class BrandingController extends Controller {
 	 */
 	public function display()
 	{
-		$branding=Branding::join('institution as i','brandings.institution_id','=','i.id')
-		->select('i.id as institution_id','brandings.id','i.name as institution_name','brandings.title')->get();
-		return view('admin::branding.brandview',compact('branding'));
+//		$branding=Branding::join('institution as i','brandings.institution_id','=','i.id')
+//		->select('i.id as institution_id','brandings.id','i.name as institution_name','brandings.title')->get();
+//		return view('admin::branding.brandview',compact('branding'));
+		$branding = Branding::join('institution as i', 'brandings.institution_id', '=', 'i.id')
+			->select('i.id as institution_id', 'brandings.id', 'i.name as institution_name')->get();
+		return view('admin::branding.brandview', compact('branding'));
+
 	}
 
 	/**
@@ -106,12 +153,14 @@ class BrandingController extends Controller {
 	 * @return Response
 	 */
 	public function edit($id)
-	{	
-			$InstitutionObj = new Institution();
+	{
+		$InstitutionObj = new Institution();
 		$inst_arr = $InstitutionObj->getInstitutions();
-			$branding=Branding::find($id);
-		
-		return view('admin::branding.brandedit',compact('branding','inst_arr'));
+		$branding=Branding::find($id);
+		$branding = Branding::find($id);
+//dd($branding->institution_id);
+		return view('admin::branding.brandedit', compact('branding', 'inst_arr'));
+
 	}
 
 	/**
@@ -121,28 +170,34 @@ class BrandingController extends Controller {
 	 * @return Response
 	 */
 	public function update($id)
-	{		
-		
-		$branding=Branding::join('institution as i','brandings.institution_id','=','i.id')
-		->select('i.id as institution_id','brandings.id','i.name as institution_name','brandings.title')->get();
-		$post=Input::all();
+	{
+
+		$branding = Branding::join('institution as i', 'brandings.institution_id', '=', 'i.id')
+			->select('i.id as institution_id', 'brandings.id', 'i.name as institution_name')->get();
+		$post = Input::all();
+
 		unset($post['_token']);
-		$record = Branding::where('id',$id)->update([
-				'title'=>$post['title'],
-			/*'filepath'=>$post['filepath'],*/
-			'header_bg_color'=>$post['hbc'],
-			'header_text_color'=>$post['htc'],
-			'box_header_bg_color'=>$post['bhbc'],
-			'box_header_text_color'=>$post['bhtc'],
-			'box_text_color'=>$post['btc'],
-			'button_color'=>$post['bc'],
-			'institution_id'=> $post['institution_id'],
-	
-				
+		$record = Branding::where('id', $id)->update([
+
+
+			//'filepath'=>$post['filepath'],
+			'institution_id' => $post['institution_id'],
+			//'title' => $post['title'],
+			'header_bg_color' => $post['hbcolor'],
+			'header_text_color' => $post['headertc'],
+			'box_header_bg_color' => $post['boxhbc'],
+			'box_header_text_color' => $post['boxhtcolor'],
+			'box_text_color' => $post['btextc'],
+			'button_color' => $post['buttonc'],
+
+
+
 		]);
-		
-		/*return Redirect::to('user/branding/brandview');*/
-		return view('admin::branding.brandview',compact('branding','record'));
+
+		//dd($record);
+		/*return redirect::to('brandview', compact('branding', 'record'));*/
+		return Redirect::route('brandview');
+
 	}
 
 	/**
@@ -154,11 +209,14 @@ class BrandingController extends Controller {
 
 	public function destroy($id)
 	{
-		
-		Branding::find($id)->delete();
-		$branding=Branding::join('institution as i','brandings.institution_id','=','i.id')
-		->select('i.id as institution_id','brandings.id','i.name as institution_name','brandings.title')->get();
-		return view('admin::branding.brandview',compact('branding'));
+
+		$branding = Branding::join('institution as i', 'brandings.institution_id', '=', 'i.id')
+			->select('i.id as institution_id', 'brandings.id', 'i.name as institution_name')->get();
+		if(!$branding == null) {
+			Branding::find($id)->delete();
+			return Redirect::route('brandview');
+		}
+
 	}
 
 }
